change folder  structure
